ci_cd/development_server/requirements.txt
keras
tensorflow==2.2.0
future
numpy
requests
matplotlib
PyGithub
scikit-learn
pandas
joblib
catboost